backend-django/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
PyMySQL==1.1.1
python-dotenv==1.0.1
